+++ perseo-quality/docs/source/changelog.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+v1.0.1
+------
+
+**Other Changes**
+
+- Added a centralized dedicated logger for the whole package
+- Changed few Log ERROR statements to WARNING
+
 v1.0.0
 ------